% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac.R
\name{clamulet}
\alias{clamulet}
\title{clamulet}
\usage{
clamulet(
  x,
  artificialDoublets = NULL,
  iter = 2,
  maxN = 500,
  nfeatures = 25,
  max_depth = 5,
  threshold = 0.75,
  returnAll = FALSE,
  ...
)
}
\arguments{
\item{x}{The path to a fragment file (see \code{\link{getFragmentOverlaps}}
for performance/memory-related guidelines)}

\item{artificialDoublets}{The number of artificial doublets to generate}

\item{iter}{The number of learning iterations (should be 1 to)}

\item{maxN}{The maximum number of regions per cell to consider to establish
windows for meta-features}

\item{nfeatures}{The number of meta-features to consider}

\item{max_depth}{The maximum tree depth}

\item{threshold}{The score threshold used during iterations}

\item{returnAll}{Logical; whether to return data also for artificial doublets}

\item{...}{Arguments passed to \code{\link{getFragmentOverlaps}}}
}
\value{
A data.frame
}
\description{
Classification-powered Amulet-like method
}
