% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{clusterStickiness}
\alias{clusterStickiness}
\title{clusterStickiness}
\usage{
clusterStickiness(
  x,
  type = c("quasibinomial", "nbinom1", "binomial", "poisson"),
  inclDiff = FALSE
)
}
\arguments{
\item{x}{A table of double statistics, or a SingleCellExperiment on which
\link{scDblFinder} was run.}

\item{type}{The type of test to use (quasibinomial recommended).}

\item{inclDiff}{Logical; whether to include the difficulty in the model. If NULL,
will be used only if there is a significant trend with the enrichment.}
}
\value{
A table of test results for each cluster.
}
\description{
Tests for enrichment of doublets created from each cluster (i.e. cluster's
stickiness). Only applicable with >=4 clusters.
}
\examples{
sce <- mockDoubletSCE(rep(200,5))
sce <- scDblFinder(sce, artificialDoublets=500)
clusterStickiness(sce)
}
