% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\alias{doubletsComparison}
\title{Comparison results}
\value{
a list.
}
\description{
Results of the comparison of doublet detection methods. Each element
of the list is a dataset, itself a list with scores for the different
methods as well as the demuxlet SNP-based doublet calls 
(`demuxlet_cls`).
}
