% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dblFinderWrapper}
\alias{dblFinderWrapper}
\title{dblFinderWrapper}
\usage{
dblFinderWrapper(sce, doublet.formation.rate = 0.025, dims = 1:10)
}
\arguments{
\item{sce}{An object of class `SingleCellExperiment`}

\item{doublet.formation.rate}{The expected doublet rate, default 0.025}

\item{dims}{The dimensions ot use, default 1:10}
}
\value{
The updated `sce` object, with the colData field `DF.score`
}
\description{
A wrapper around \url{https://github.com/chris-mcginnis-ucsf/DoubletFinder}[DoubletFinder],
used for comparison. Requires the `DoubletFinder` and `Seurat` version 3 
packages to be installed.
}
