% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{doubletPairwiseEnrichment}
\alias{doubletPairwiseEnrichment}
\title{doubletPairwiseEnrichment}
\usage{
doubletPairwiseEnrichment(
  x,
  lower.tail = FALSE,
  sampleWise = FALSE,
  type = c("poisson", "binomial", "chisq", "nbinom1")
)
}
\arguments{
\item{x}{A table of double statistics, or a SingleCellExperiment on which scDblFinder
was run.}

\item{lower.tail}{Logical; defaults to FALSE to test enrichment (instead of depletion).}

\item{sampleWise}{Logical; whether to perform tests sample-wise in multi-sample
datasets. If FALSE (default), will aggregate counts before testing.}

\item{type}{Type of test to use.}
}
\value{
A table of significances for each combination.
}
\description{
doubletPairwiseEnrichment
}
\examples{
sce <- mockDoubletSCE()
sce <- scDblFinder(sce, artificialDoublets=500)
doubletPairwiseEnrichment(sce)
}
