% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overcluster.R
\name{fastClust}
\alias{fastClust}
\title{fastClust}
\usage{
fastClust(sce, nfeatures = 1000, k = 10, dims = 20,
  graph.type = c("snn", "knn"), method = c("louvain", "fast_greedy",
  "overcluster"), BPPARAM = BiocParallel::SerialParam(), ...)
}
\arguments{
\item{sce}{An object of class `SingleCellExperiment`}

\item{nfeatures}{For the PCA}

\item{k}{number of nearest neighbors}

\item{dims}{number of PCA dimensions}

\item{graph.type}{either snn or knn}

\item{method}{Either 'louvain' or 'fast_greedy'}

\item{BPPARAM}{Passed to scran for KNN/SNN graph generation}

\item{...}{passed to `overcluster`}
}
\value{
The `SingleCellExperiment` object with an additional colData column
scDblFinder.clusters
}
\description{
fastClust
}
