% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{fastcluster}
\alias{fastcluster}
\title{fastcluster}
\usage{
fastcluster(
  x,
  k = NULL,
  rdname = "PCA",
  nstart = 2,
  iter.max = 20,
  ndims = NULL,
  nfeatures = 1000
)
}
\arguments{
\item{x}{An object of class SCE}

\item{k}{The number of k-means clusters to use in the primary step (should 
be much higher than the number of expected clusters). Defaults to 1/10th of 
the number of cells with a maximum of 3000.}

\item{rdname}{The name of the dimensionality reduction to use.}

\item{nstart}{Number of starts for k-means clustering}

\item{iter.max}{Number of iterations for k-means clustering}

\item{ndims}{Number of dimensions to use}

\item{nfeatures}{Number of features to use if doing PCA}
}
\value{
A vector of cluster labels
}
\description{
Performs a fast two-step clustering: first clusters using k-means with a very
large k, then uses louvain clustering of the k cluster averages and reports
back the cluster labels.
}
