% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyDoublets.R
\name{identifyDoubletOrigins}
\alias{identifyDoubletOrigins}
\title{identifyDoubletOrigins}
\usage{
identifyDoubletOrigins(
  sce,
  clusters,
  samples = NULL,
  doublets = NULL,
  balance = TRUE,
  nArtificial = NULL,
  verbose = TRUE,
  xgb.param = list(booster = "gbtree", objective = "multi:softprob", eval_metric =
    "mlogloss", subsample = 0.5, colsample_bytree = 0.4, eta = 0.5, lambda = 100, alpha =
    1),
  max_rounds = 300,
  nthread = 1
)
}
\arguments{
\item{sce}{A SingleCellExperiment object with a 'counts' assay.}

\item{clusters}{A vector of cluster labels for each column of `sce`, or the 
name of a colData column of `sce` containing such labels.}

\item{samples}{An optional vector of sample labels for each column of `sce`,
or the name of a colData column of `sce` containing such labels. If 
provided, artificial doublets will be generated within-sample (but the 
classifier trained across samples).}

\item{doublets}{The doublets to be identified. If NULL, doublets will be 
taken from `sce$scDblFinder.class` if present; if not, only training is
performed.}

\item{balance}{Logical; whether to balance doublet types (default TRUE).}

\item{nArtificial}{Number of artificial doublets. If omitted, 100 per 
cluster combination will be used, up to a maximum of 20000.}

\item{verbose}{Logical; whether to output progress messages.}

\item{xgb.param}{A named list of parameters passed to xgboost.}

\item{max_rounds}{The maximum training round during cross-validation.}

\item{nthread}{The number of threads.}
}
\value{
A list with:
   - model : the xgboost model
   - train_contigency : the contingency matrix on the training data
   - predictions : the per-class probabilities on `doublets` (if given)
   - calls : the origin calls on `doublets` (if given)
   - features : the ordered features (i.e. genes) needed to run the model.
}
\description{
Trains a classifier based on artificial doublets to identify the origins of 
doublets.
}
\examples{
# we generate a random dataset
sce <- mockDoubletSCE(ncells = c(20,30,40), ngenes = 500)
# to have the example run fast, we set a low number of artificial doublets 
# and a low maximum learning rounds
res <- identifyDoubletOrigins(sce, "cluster", nArtificial=100, max_rounds=10)
# if desired, we could then re-run the same classifier on a new sample using
# predictDoubletOrigins()
}
