% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{overcluster}
\alias{overcluster}
\title{overcluster}
\usage{
overcluster(x, min.size = 50, max.size = NULL, rdname = "PCA", ...)
}
\arguments{
\item{x}{A numeric matrix, with entities (e.g. cells) as columns and features
(e.g. genes) as rows. Alternatively, an object of class `igraph`.}

\item{min.size}{The minimum cluster size (applies after splitting, and hence 
overrides `max.size`)}

\item{max.size}{The maximum cluster size. If omitted, will be calculated on 
the basis of the population size and initial number of clusters.}

\item{rdname}{Optional name of the reduced dimension to use.}
}
\value{
A vector of cluster labels.
}
\description{
This function deliberately overclusters based on the desired range of cluster
size. It uses iterative calls to `igraph::cluster_fast_greedy` until no 
cluster is above the size limits, and merges clusters that are too small.
}
\examples{
m <- t(sapply( seq(from=0, to=5, length.out=50), 
               FUN=function(x) rpois(50,x) ) )
cc <- suppressWarnings(overcluster(m,min.size=5))
table(cc)

}
