% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plotROCs}
\alias{plotROCs}
\title{plotROCs}
\usage{
plotROCs(scores, truth)
}
\arguments{
\item{scores}{A data.frame with the different types of scores as columns.}

\item{truth}{A vector of the true class corresponding to each row of `scores`}
}
\value{
a ggplot
}
\description{
Plot ROC curves for given scores.
}
