% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotThresholds}
\alias{plotThresholds}
\title{plotThresholds}
\usage{
plotThresholds(d, ths = (0:100)/100, dbr = NULL, dbr.sd = NULL, do.plot = TRUE)
}
\arguments{
\item{d}{A data.frame of cell properties, with each row representing a cell,
as produced by `scDblFinder(..., returnType="table")`.}

\item{ths}{A vector of thresholds between 0 and 1 at which to plot values.}

\item{dbr}{The expected (mean) doublet rate.}

\item{dbr.sd}{The standard deviation of the doublet rate, representing the
uncertainty in the estimate.}

\item{do.plot}{Logical; whether to plot the data (otherwise will return the
underlying data.frame).}
}
\value{
A ggplot, or a data.frame if `do.plot==FALSE`.
}
\description{
Plots scores used for thresholding.
}
