% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyDoublets.R
\name{predictDoubletOrigins}
\alias{predictDoubletOrigins}
\title{predictDoubletOrigins : run an origins classifier on new cells}
\usage{
predictDoubletOrigins(model, doublets, ret = c("call", "probs"))
}
\arguments{
\item{model}{The output of \code{\link{identifyDoubletOrigins}}.}

\item{doublets}{A `SingleCellExperiment` or counts matrix of doublets.}

\item{ret}{Either 'call' (origin call, default) or 'probs' (per-class 
probabilities).}
}
\value{
Either a factor (for `ret="call"`) or a matrix of probabilities.
}
\description{
predictDoubletOrigins : run an origins classifier on new cells
}
\examples{
# we generate a random dataset
sce <- mockDoubletSCE(ncells = c(20,30,40), ngenes = 500)
# to have the example run fast, we set a low number of artificial doublets 
# and a low maximum learning rounds
clf <- identifyDoubletOrigins(sce, "cluster", nArtificial=100, max_rounds=10)
# we run on a new set of doublets:
isDoublet <- which(sce$type=="doublet")
res <- predictDoubletOrigins(clf, sce[,isDoublet])
table(true=sce$origin[isDoublet], res)
}
