% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{propHomotypic}
\alias{propHomotypic}
\title{propHomotypic}
\usage{
propHomotypic(clusters)
}
\arguments{
\item{clusters}{A vector of cluster labels}
}
\value{
A numeric value between 0 and 1.
}
\description{
Computes the proportion of pairs expected to be made of elements from the 
same cluster.
}
\examples{
clusters <- sample(LETTERS[1:5], 100, replace=TRUE)
propHomotypic(clusters)
}
