% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rankTrans}
\alias{rankTrans}
\title{rankTrans}
\usage{
rankTrans(x)
}
\arguments{
\item{x}{A matrix, with samples/cells as columns and genes/features as rows.}
}
\value{
rank-transformed x.
}
\description{
A dense rank transformation that preserves 0 and rank step size in the 
presence of many ties.
}
\examples{
m <- t(sapply( seq(from=0, to=5, length.out=30), 
               FUN=function(x) rpois(30,x) ) )
m2 <- rankTrans(m)

}
