% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{scdsWrapper}
\alias{scdsWrapper}
\title{scdsWrapper}
\usage{
scdsWrapper(sce)
}
\arguments{
\item{sce}{An object of class `SingleCellExperiment`}
}
\value{
The updated `sce` object, with the colData field `hybrid_score`
}
\description{
A wrapper around scds's hybrid score, used for comparison (requires the 
`scds` package to be installed). This function is not exported, and is there
solely as a record of the specific way the comparisons where made.
}
