% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_processing.R
\name{TFIDF}
\alias{TFIDF}
\title{TFIDF}
\usage{
TFIDF(x, sf = 10000)
}
\arguments{
\item{x}{The matrix of occurrences}

\item{sf}{Scaling factor}
}
\value{
An array of same dimensions as `x`
}
\description{
The Term Frequency - Inverse Document Frequency (TF-IDF) normalization, as
implemented in Stuart & Butler et al. 2019.
}
\examples{
m <- matrix(rpois(500,1),nrow=50)
m <- TFIDF(m)
}
