% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac_processing.R
\name{aggregateFeatures}
\alias{aggregateFeatures}
\title{aggregateFeatures}
\usage{
aggregateFeatures(
  x,
  dims.use = seq(2L, 12L),
  k = 1000,
  num_init = 3,
  use.mbk = NULL,
  use.subset = 5000,
  use.TFIDF = TRUE,
  ...
)
}
\arguments{
\item{x}{A integer/numeric (sparse) matrix, or a `SingleCellExperiment`
including a `counts` assay.}

\item{dims.use}{The PCA dimensions to use for clustering rows.}

\item{k}{The approximate number of meta-features desired}

\item{num_init}{The number of initializations used for k-means clustering.}

\item{use.mbk}{Logical; whether to use minibatch k-means (see
\code{\link[mbkmeans]{mbkmeans}}). If NULL, the minibatch approach will be
used if there are more than 30000 features.}

\item{use.subset}{How many cells (columns) to use to cluster the features.}

\item{use.TFIDF}{Logical; whether to use \link{TFIDF} normalization (instead
of standard normalization) to assess the similarity between features.
similarity}

\item{...}{Passed to \code{\link[mbkmeans]{mbkmeans}}. Can for instance be
used to pass the `BPPARAM` argument for multithreading.}
}
\value{
An aggregated version of `x` (either an array or a
`SingleCellExperiment`, depending on the input).
}
\description{
Aggregates similar features (rows).
}
