% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletThresholding.R
\name{doubletThresholding}
\alias{doubletThresholding}
\title{doubletThresholding}
\usage{
doubletThresholding(d, dbr = 0.025, dbr.sd = 0.015, local = TRUE)
}
\arguments{
\item{d}{A data.frame of properties of real and artificial cells, as produced
by `scDblFinder(returnType="table")`.}

\item{dbr}{The expected (mean) doublet rate.}

\item{dbr.sd}{The standard deviation of the doublet rate, representing the 
uncertainty in the estimate.}

\item{local}{Logical; whether to use local calibration (experimental 
feature!)}
}
\value{
A scaler indicating the decided threshold.
}
\description{
Sets the doublet scores threshold; typically called by 
\code{\link[scDblFinder]{scDblFinder}}.
}
\examples{
sce <- mockDoubletSCE()
d <- scDblFinder(sce, verbose=FALSE, returnType="table")
th <- doubletThresholding(d, dbr=0.05)
th$th

}
