% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletThresholding.R
\name{doubletThresholding}
\alias{doubletThresholding}
\title{doubletThresholding}
\usage{
doubletThresholding(scores, celltypes, dbr = 0.021, dbr.sd = 0.01,
  clusters = NULL, do.plot = TRUE)
}
\arguments{
\item{scores}{A vector of the doublet score for each cell (real and artificial); can be
anything ranging from 0 to 1, with higher scores indicating higher change of being a 
doublet.}

\item{celltypes}{A vector of the same length as `scores` indicating, for each cell,
whether it is a 'real' cell or an 'artificial' doublet. Missing values not allowed.}

\item{dbr}{The expected (mean) doublet rate, defaults to 0.021 on the basis of the 
mixology 10x datasets demuxlet results, where the proportion of demuxlet doublet is
estimated respectively at 0.012 and 0.029 (after accounting adding expected homotypic 
doublets).}

\item{dbr.sd}{The deviation of the doublet rate, representing the uncertainty in the 
estimate. Effectively the scale of a Cauchy distribution. Defaults to 0.01.}

\item{clusters}{Optional vector of cluster assignment for each (real) cell, used for
homotypic doublet correction.}

\item{do.plot}{Logical; whether to plot the thresholding data (default TRUE).}
}
\value{
A scaler indicating the decided threshold.
}
\description{
doubletThresholding
}
