% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletThresholding.R
\name{doubletThresholding}
\alias{doubletThresholding}
\title{doubletThresholding}
\usage{
doubletThresholding(scores, celltypes, clusters = NULL, dbr = 0.025,
  dbr.sd = 0.02, prop.fullyRandom = 0, do.plot = TRUE)
}
\arguments{
\item{scores}{A vector of the doublet score for each cell (real and artificial); can be
anything ranging from 0 to 1, with higher scores indicating higher change of being a 
doublet.}

\item{celltypes}{A vector of the same length as `scores` indicating, for each cell,
whether it is a 'real' cell or an 'artificial' doublet. Missing values not allowed.}

\item{clusters}{Optional vector of cluster assignment for each (real) cell, used for
homotypic doublet correction.}

\item{dbr}{The expected (mean) doublet rate.}

\item{dbr.sd}{The standard deviation of the doublet rate, representing the uncertainty
in the estimate.}

\item{prop.fullyRandom}{The proportion of artificical doublets that are fully random,
used for homotypy correction. Default 0.25 (the default value in 
`getArtificialDoublets`). Ignored if `clusters=NULL`}

\item{do.plot}{Logical; whether to plot the thresholding data (default TRUE).}
}
\value{
A scaler indicating the decided threshold.
}
\description{
doubletThresholding
}
