% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArtificialDoublets.R
\name{getArtificialDoublets}
\alias{getArtificialDoublets}
\title{getArtificialDoublets}
\usage{
getArtificialDoublets(
  x,
  n = 3000,
  clusters = NULL,
  n.meta.cells = 1,
  meta.triplets = TRUE
)
}
\arguments{
\item{x}{A count matrix, with features as rows and cells as columns.}

\item{n}{The approximate number of doublet to generate (default 3000).}

\item{clusters}{The optional clusters labels to use to build cross-cluster 
doublets.}

\item{n.meta.cells}{The number of meta-cell per cluster to create. If given,
additional doublets will be created from cluster meta-cells.}

\item{meta.triplets}{Logical; whether to create triplets from meta cells. 
Ignored if `clusters` is missing.}

\item{prop.fullyRandom}{The proportion of the created doublets that are fully
random (default 0); the rest will be doublets created across clusters. 
Ignored if `clusters` is NULL.}
}
\value{
A count matrix for artificial doublets.
}
\description{
Create expression profiles of random artificial doublets.
}
\examples{
m <- t(sapply( seq(from=0, to=5, length.out=50), 
               FUN=function(x) rpois(30,x) ) )
doublets <- getArtificialDoublets(m, 30)

}
