% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArtificialDoublets.R
\name{getCellPairs}
\alias{getCellPairs}
\title{getCellPairs}
\usage{
getCellPairs(
  clusters,
  n = 1000,
  ls = NULL,
  q = c(0.1, 0.9),
  selMode = "proportional",
  soft.min = 5,
  ...
)
}
\arguments{
\item{n}{The number of cell pairs to obtain}

\item{...}{Further arguments, for instance the `k` vector of precluster
labels if `x` is a metacell graph.}

\item{x}{A vector of cluster labels for each cell, or a list containing
metacells and graph}
}
\value{
A data.frame with the columns
}
\description{
Given a vector of cluster labels, returns pairs of cross-cluster cells
}
\examples{
# create random labels
x <- sample(head(LETTERS), 100, replace=TRUE)
getCellPairs(x, n=6)
}
