% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rankTrans}
\alias{rankTrans}
\title{rankTrans}
\usage{
rankTrans(x)
}
\arguments{
\item{x}{A matrix, with samples/cells as columns and genes/features as rows.}
}
\value{
rank-transformed x.
}
\description{
A dense rank transformation that preserves 0 and rank step size in the presence of many
ties.
}
