% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletFinder.R
\name{scDblFinder}
\alias{scDblFinder}
\title{scDblFinder}
\usage{
scDblFinder(sce, artificialDoublets = NULL, clusters = NULL,
  minClusSize = 50, maxClusSize = ceiling(ncol(sce)/20), d = 10,
  dbr = 0.025, dbr.sd = 0.015, k = 5, graph.type = c("knn", "snn"),
  fullTable = FALSE, trans = "scran", verbose = TRUE,
  BPPARAM = SerialParam())
}
\arguments{
\item{sce}{An object of class `SingleCellExperiment`}

\item{artificialDoublets}{The approximate number of artificial doublets to create. If
NULL, will be the maximum of the number of cells or `3*nbClusters^2`.}

\item{clusters}{The optional cluster assignments (if ommitted, will run `quickCluster`).
This is used to make doublets more efficiently.}

\item{minClusSize}{The minimum cluster size for `quickCluster`/`overcluster` 
(default 50); ignored if `clusters` is given.}

\item{maxClusSize}{The maximum cluster size for `overcluster`. Ignored if `clusters` is
given. If NULL or NA, clustering will be performed using `quickCluster`, otherwise via
`overcluster`. The default value is `ncol(sce)/20`.}

\item{d}{The number of dimensions used to build the KNN network (default 10)}

\item{dbr}{The expected doublet rate. Defaults to 0.025 the basis of the mixology 10x 
datasets demuxlet results, where the proportion of demuxlet doublet is estimated 
respectively at 0.012 and 0.029 (after accounting adding expected homotypic doublets). 
However, note that this value might not be appropriate for all datasets, especially 
older 10x datasets, where the proportion of doublets can go up to 0.1.}

\item{dbr.sd}{The standard deviation of the doublet rate, defaults to 0.02.}

\item{k}{Number of nearest neighbors (for KNN graph).}

\item{graph.type}{Either 'snn' or 'knn' (default).}

\item{fullTable}{Logical; whether to return the full table including artificial 
doublets (default FALSE), rather than the table for real cells only.}

\item{trans}{}

\item{verbose}{Logical; whether to print messages and the thresholding plot (default
TRUE).}

\item{BPPARAM}{Passed to scran; doesn't work so well...}
}
\value{
A data.frame including, for each cells (rows, in the order in which they appear
in the `sce` object), the number of neighbors considered, the number of these neighbors
 that are from artificial doublets, the ratio of neighbors that are from artificial 
 doublets (which represents a quantitative doublet the score), the enrichment in 
 artificial doublets, and the classification (after applying the threshold).
}
\description{
Identification of doublets in single-cell RNAseq directly from counts using 
overclustering-based generation of artifical doublets.
}
