% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDblFinder.R
\name{scDblFinder}
\alias{scDblFinder}
\title{scDblFinder}
\usage{
scDblFinder(sce, artificialDoublets = NULL, clusters = NULL,
  samples = NULL, minClusSize = min(50, ncol(sce)/5),
  maxClusSize = NULL, d = 10, dbr = NULL, dbr.sd = 0.015, k = 5,
  graph.type = c("knn", "snn"), fullTable = FALSE, trans = c("scran",
  "rankTrans", "none", "lognorm"), verbose = is.null(samples),
  BPPARAM = SerialParam())
}
\arguments{
\item{sce}{A \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}}

\item{artificialDoublets}{The approximate number of artificial doublets to 
create. If NULL, will be the maximum of the number of cells or 
`3*nbClusters^2`.}

\item{clusters}{The optional cluster assignments (if ommitted, will run 
`quickCluster`). This is used to make doublets more efficiently.}

\item{samples}{A vector of the same length as cells (or the name of a column of 
`colData(sce)`), indicating to which sample each cell belongs. Here, a sample
is understood as being processed independently. If ommitted, doublets will be
searched for with all cells together. If given, doublets will be searched for 
independently for each sample, which is preferable if they represent different
captures.}

\item{minClusSize}{The minimum cluster size for `quickCluster`/`overcluster` 
(default 50); ignored if `clusters` is given.}

\item{maxClusSize}{The maximum cluster size for `overcluster`. Ignored if 
`clusters` is given. If NA, clustering will be performed using `quickCluster`,
otherwise via `overcluster`. If missing, the default value will be estimated 
by `overcluster`.}

\item{d}{The number of dimensions used to build the KNN network (default 10)}

\item{dbr}{The expected doublet rate. By default this is assumed to be 
0.1*nbCells percent, which is appropriate for 10x datasets.}

\item{dbr.sd}{The standard deviation of the doublet rate, defaults to 0.015.}

\item{k}{Number of nearest neighbors (for KNN graph).}

\item{graph.type}{Either 'snn' or 'knn' (default).}

\item{fullTable}{Logical; whether to return the full table including artificial 
doublets (default FALSE), rather than the table for real cells only.}

\item{trans}{The transformation to use before computing the KNN network. The 
default, `scran::scaledColRanks`, gave the best result in our hands.}

\item{verbose}{Logical; whether to print messages and the thresholding plot.}

\item{BPPARAM}{Used for multithreading when splitting by samples (i.e. when 
`samples!=NULL`); otherwise passed to scran (which doesn't work so well)...}
}
\value{
The `sce` object with the following additional colData columns: 
`scDblFinder.neighbors` (number of neighbors considered), `scDblFinder.ratio`
(ratio of aritifical doublets among neighbors), and `scDblFinder.class` 
(whether the cell is called as 'doublet' or 'singlet'). Alternatively, if 
`fullTable=TRUE`, a data.frame will be returned with information about real 
and artificial cells.
}
\description{
Identification of doublets in single-cell RNAseq directly from counts using 
overclustering-based generation of artifical doublets.
}
\examples{
library(SingleCellExperiment)
m <- t(sapply( seq(from=0, to=5, length.out=50), 
               FUN=function(x) rpois(50,x) ) )
sce <- SingleCellExperiment( list(counts=m) )
sce <- scDblFinder(sce, verbose=FALSE)
table(sce$scDblFinder.class)

}
